+++ decidim-assemblies/spec/system/admin/admin_manages_assemblies_with_parent_selector_spec.rb
@@ -1,0 +1,144 @@
+# frozen_string_literal: true
+
+require "spec_helper"
+
+describe "Admin manages assemblies with parent selector", type: :system do
+  include_context "when admin administrating an assembly"
+
+  let(:image1_filename) { "city.jpeg" }
+  let(:image1_path) { Decidim::Dev.asset(image1_filename) }
+  let(:image2_filename) { "city2.jpeg" }
+  let(:image2_path) { Decidim::Dev.asset(image2_filename) }
+  let!(:set_all_assemblies) { Decidim::Assemblies::OrganizationAssemblies.new(organization).query }
+
+  context "when params[:parent_id] not present" do
+    before do
+      switch_to_host(organization.host)
+      login_as user, scope: :user
+      visit decidim_admin_assemblies.assemblies_path
+      click_link "New assembly"
+    end
+
+    let!(:params) do
+      {
+        parent_id: nil
+      }
+    end
+
+    it "creates a new assembly" do
+      within ".new_assembly" do
+        fill_in_i18n(
+          :assembly_title,
+          "#assembly-title-tabs",
+          en: "My assembly",
+          es: "Mi proceso participativo",
+          ca: "El meu procés participatiu"
+        )
+        fill_in_i18n(
+          :assembly_subtitle,
+          "#assembly-subtitle-tabs",
+          en: "Subtitle",
+          es: "Subtítulo",
+          ca: "Subtítol"
+        )
+        fill_in_i18n_editor(
+          :assembly_short_description,
+          "#assembly-short_description-tabs",
+          en: "Short description",
+          es: "Descripción corta",
+          ca: "Descripció curta"
+        )
+        fill_in_i18n_editor(
+          :assembly_description,
+          "#assembly-description-tabs",
+          en: "A longer description",
+          es: "Descripción más larga",
+          ca: "Descripció més llarga"
+        )
+
+        select set_all_assemblies.first.title["en"], from: :assembly_parent_id
+
+        fill_in :assembly_slug, with: "slug"
+        fill_in :assembly_hashtag, with: "#hashtag"
+        attach_file :assembly_hero_image, image1_path
+        attach_file :assembly_banner_image, image2_path
+
+        find("*[type=submit]").click
+      end
+
+      expect(page).to have_admin_callout("successfully")
+
+      within ".container" do
+        expect(page).to have_current_path decidim_admin_assemblies.assemblies_path(parent_id: set_all_assemblies.first&.id)
+        expect(page).to have_content("My assembly")
+      end
+    end
+  end
+
+  context "when managing child assemblies" do
+    let!(:parent_assembly) { create :assembly, organization: organization }
+    let!(:child_assembly) { create :assembly, organization: organization, parent: parent_assembly }
+    let(:assembly) { child_assembly }
+    let(:image1_filename) { "city.jpeg" }
+    let(:image1_path) { Decidim::Dev.asset(image1_filename) }
+    let(:image2_filename) { "city2.jpeg" }
+    let(:image2_path) { Decidim::Dev.asset(image2_filename) }
+
+    before do
+      switch_to_host(organization.host)
+      login_as user, scope: :user
+      visit decidim_admin_assemblies.assemblies_path
+      within find("tr", text: translated(parent_assembly.title)) do
+        click_link "Assemblies"
+      end
+      click_link "New assembly"
+    end
+
+    it "creates a new assembly" do
+      within ".new_assembly" do
+        fill_in_i18n(
+          :assembly_title,
+          "#assembly-title-tabs",
+          en: "My assembly",
+          es: "Mi proceso participativo",
+          ca: "El meu procés participatiu"
+        )
+        fill_in_i18n(
+          :assembly_subtitle,
+          "#assembly-subtitle-tabs",
+          en: "Subtitle",
+          es: "Subtítulo",
+          ca: "Subtítol"
+        )
+        fill_in_i18n_editor(
+          :assembly_short_description,
+          "#assembly-short_description-tabs",
+          en: "Short description",
+          es: "Descripción corta",
+          ca: "Descripció curta"
+        )
+        fill_in_i18n_editor(
+          :assembly_description,
+          "#assembly-description-tabs",
+          en: "A longer description",
+          es: "Descripción más larga",
+          ca: "Descripció més llarga"
+        )
+
+        fill_in :assembly_slug, with: "slug"
+        fill_in :assembly_hashtag, with: "#hashtag"
+        attach_file :assembly_hero_image, image1_path
+        attach_file :assembly_banner_image, image2_path
+
+        find("*[type=submit]").click
+      end
+
+      expect(page).to have_admin_callout("successfully")
+
+      within ".container" do
+        expect(page).to have_current_path decidim_admin_assemblies.assemblies_path(parent_id: parent_assembly&.id)
+        expect(page).to have_content("My assembly")
+      end
+    end
+  end
+end